requirements.txt
# Core
streamlit>=1.35.0
python-dotenv>=1.0.0

# Google Gemini
google-genai>=1.0.0

# Vector search
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0

# Document extraction
pdfplumber>=0.11.0          # PDF text + table extraction
pymupdf>=1.24.0             # PDF image extraction (fitz)
python-docx>=1.1.0          # Word documents
python-pptx>=0.6.23         # PowerPoint
openpyxl>=3.1.0             # Excel

# Image handling (transitive but explicit is better)
Pillow>=10.0.0

# Numerical
numpy>=1.26.0